Just committing so that I can pull from my desktop

+++ SimpleCalculator/InputConverter.cs
@@ -4,10 +4,10 @@ namespace SimpleCalculator
 {
     public class InputConverter
     {
-        public static bool ConvertInputToNumeric(string argTextInput, out int result)
+        public static bool ConvertInputToNumeric(string argTextInput, out double result)
         {
 
-            bool success = int.TryParse(argTextInput, out int localResult);
+            bool success = double.TryParse(argTextInput, out double localResult);
 
             result = localResult;
             return success;
